requirements.txt
agentevals==0.0.9
annotated-types==0.7.0
anthropic==0.59.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
dydantic==0.0.8
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
langchain==0.3.27
langchain-anthropic==0.3.17
langchain-core==0.3.72
langchain-mcp-adapters==0.1.9
langchain-openai==0.3.28
#langchain-aws==0.1.16
langchain-redis==0.2.3
langchain-text-splitters==0.3.9
langgraph==0.5.4
langgraph-checkpoint==2.1.1
langgraph-checkpoint-postgres==2.0.23
langgraph-checkpoint-redis==0.0.8
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.74
langgraph-supervisor==0.0.28
langgraph-swarm==0.0.13
langmem==0.0.28
langsmith==0.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.12.2
mdurl==0.1.2
ml_dtypes==0.5.1
numpy==2.3.2
openai==1.97.1
openevals==0.1.0
orjson==3.11.1
ormsgpack==1.10.0
packaging==25.0
pgvector==0.4.1
ply==3.11
psycopg==3.2.9
psycopg-pool==3.2.6
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
python-multipart==0.0.20
python-ulid==3.0.0
PyYAML==6.0.2
redis==6.2.0
redisvl==0.8.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.26.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sse-starlette==2.4.1
starlette==0.47.2
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
trustcall==0.0.39
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
xxhash==3.5.0
zstandard==0.23.0
langfuse==3.2.3
azure-storage-blob==12.23.1